% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregates results from the clustering and merging step.}
\usage{
aggregateData(dataInput, labelsInput)
}
\arguments{
\item{dataInput}{File name of processed data in tab delimited format. Rows are events and columns are features measured.}

\item{labelsInput}{File name of output from clustering and merging step.}
}
\value{
The aggregated data of \code{dataInput}, with average signal levels for all clusters and sample combinations.
}
\description{
Aggregates results from the clustering and merging step.
}

