% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateClades.R
\name{annotateClades}
\alias{annotateClades}
\title{Creates annotation matrix for the clades in aggregated format. The matrix contains average signals of each dimension for each clade in each sample}
\usage{
annotateClades(sampleIDs, topHubs)
}
\arguments{
\item{sampleIDs}{sampleID vector}

\item{topHubs}{number of top ranked genes to output for annotation; annotation is a concatenated list of top ranked genes.}
}
\value{
Annotated clade matrix
}
\description{
Creates annotation matrix for the clades in aggregated format. The matrix contains average signals of each dimension for each clade in each sample
}

