% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationMatrix_withSubpopProp.R
\name{annotationMatrix_withSubpopProp}
\alias{annotationMatrix_withSubpopProp}
\title{Adds subpopulation proportion for the annotation matrix for the clades}
\usage{
annotationMatrix_withSubpopProp(aggregateMatrix_withAnnotation)
}
\arguments{
\item{aggregateMatrix_withAnnotation}{the annotated clade matrix}
}
\value{
Annotated clade matrix with subpopulation proportions
}
\description{
Adds subpopulation proportion for the annotation matrix for the clades
}

