% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAC.R
\name{BSPLeaveCenter}
\alias{BSPLeaveCenter}
\title{Finds N Leaf centers in the data}
\usage{
BSPLeaveCenter(data, N = 40, method = "dsp")
}
\arguments{
\item{data}{a n x p data matrix}

\item{N}{number of leaves centers}

\item{method}{partition method, either "dsp (discrepancy based partition)",  or "ll (bayesian sequantial partition limited-look ahead)"}
}
\value{
leafctr  N leaves centers
}
\description{
Finds N Leaf centers in the data
}

