% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterNames.R
\name{clusterNames}
\alias{clusterNames}
\title{Get the top features that differentiate the clusters based on the maximum cluster signal levels across all samples. The maximum signal levels of all features of each cluster are found, and then the maximum signal levels are normalized across clusters. Next, within each cluster, the ranks of the features are obtained based on the normalized signal levels. Top ranked features are used to annotate each cluster}
\usage{
clusterNames(aggregatedData, Num_TopFeatures = 3)
}
\arguments{
\item{aggregatedData}{Variable name of aggregated data.}

\item{Num_TopFeatures}{The number of top features in descending order of rank of importance.}
}
\value{
The annotation of clusters based on top features
}
\description{
Get the top features that differentiate the clusters based on the maximum cluster signal levels across all samples. The maximum signal levels of all features of each cluster are found, and then the maximum signal levels are normalized across clusters. Next, within each cluster, the ranks of the features are obtained based on the normalized signal levels. Top ranked features are used to annotate each cluster
}

