% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPropHeatmap.R
\name{clusterPropHeatmap}
\alias{clusterPropHeatmap}
\title{Make heatmap illustrating the cluster proportions across multiple samples. The aggregated data are first padded to assign size of 0 to missing clusters in some samples. Next, the numbers of events in each cluster in each sample are obtained. These values are normalized across samples to find the cluster proportions by samples. The higher the cluster proportion in one sample, the more specific the cluster is to that sample.}
\usage{
clusterPropHeatmap(data_agg, Colv_order)
}
\arguments{
\item{data_agg}{Variable name of aggregated data.}

\item{Colv_order}{Variable vector specifying the order of heatmap columns}
}
\value{
Plots cluster proportion heatmap plot and returns a matrix of the proportion values.
}
\description{
Make heatmap illustrating the cluster proportions across multiple samples. The aggregated data are first padded to assign size of 0 to missing clusters in some samples. Next, the numbers of events in each cluster in each sample are obtained. These values are normalized across samples to find the cluster proportions by samples. The higher the cluster proportion in one sample, the more specific the cluster is to that sample.
}

