% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{constellationPlot}
\alias{constellationPlot}
\title{Makes constellation plot, in which the centroids are clusters are embedded in the t-SNE 2D plane and the cross-sample relationships are plotted as lines connecting related sample clusters (clades).}
\usage{
constellationPlot(pacman_results, perplexity, max_iter, seed,
  plotTitle = "Constellations of Clades", nudge_x = 0.3, nudge_y = 0.3)
}
\arguments{
\item{pacman_results}{PAC-MAN analysis result matrix that contains network annotation, clade IDs and mean (centroid) clade expression levels.}

\item{perplexity}{perplexity setting for running t-SNE}

\item{max_iter}{max_iter setting for running t-SNE}

\item{seed}{set seed to make t-SNE and consetllation plot to be reproducible}

\item{plotTitle}{max_iter setting for running t-SNE}

\item{nudge_x}{nudge on x coordinate of centroid labels}

\item{nudge_y}{nudge on y coordinate of centroid labels}
}
\description{
Makes constellation plot, in which the centroids are clusters are embedded in the t-SNE 2D plane and the cross-sample relationships are plotted as lines connecting related sample clusters (clades).
}

