% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredResults_JSON.R
\name{filteredResults_JSON}
\alias{filteredResults_JSON}
\title{Obtain the input to zoomable packed circles plot. The clusters are filtered by size.}
\usage{
filteredResults_JSON(toJSON_Input, CountFilter = 1000,
  filename = "zoomablePackedCirclesInput.txt")
}
\arguments{
\item{toJSON_Input}{Variable name of the JSON input file.}

\item{CountFilter}{The minimum size of retained clusters in each sample. Set to 0 for no filtering.}

\item{filename}{The filename for zoomable packed circles plot.}
}
\value{
The JSON format input to D3 zoomable packed circles plot.
}
\description{
Obtain the input to zoomable packed circles plot. The clusters are filtered by size.
}

