% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmeasure.R
\name{fmeasure}
\alias{fmeasure}
\title{F-measure Calculation}
\usage{
fmeasure(g, t)
}
\arguments{
\item{g}{the ground truth}

\item{t}{estimated labels}
}
\value{
f  the F measure
}
\description{
Compute the F measure between the ground truth and the estimated label
}

