% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAverageSpreadOf2SubpopClades.R
\name{getAverageSpreadOf2SubpopClades}
\alias{getAverageSpreadOf2SubpopClades}
\title{Calculate the (global) average spread of subpopulations in clades with 2 subpopulations on the constellation plot.}
\usage{
getAverageSpreadOf2SubpopClades(tsneResults, pacman_results)
}
\arguments{
\item{tsneResults}{t-SNE output of clade centroids' embedding.}

\item{pacman_results}{PAC-MAN analysis result matrix that contains network annotation, clade IDs and mean (centroid) clade expression levels.}
}
\value{
Returns global average of 2-subpopulation clade spread on the constellation plot.
}
\description{
Calculate the (global) average spread of subpopulations in clades with 2 subpopulations on the constellation plot.
}

