% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepresentativeNetworks.R
\name{getRepresentativeNetworks}
\alias{getRepresentativeNetworks}
\title{Outputs representative networks for clades/subpopulations larger than a size filter (very small subpopulations are not considered in downstream analyses)}
\usage{
getRepresentativeNetworks(sampleIDs, SubpopSizeFilter, num_networkEdge)
}
\arguments{
\item{sampleIDs}{sampleID vector}

\item{SubpopSizeFilter}{the cutoff for small subpopulations. Smaller subpopulations have unstable covariance structure, so no network structure is calculated}

\item{num_networkEdge}{the number of edges to draw for each subpopulation mutual information network}
}
\description{
Outputs representative networks for clades/subpopulations larger than a size filter (very small subpopulations are not considered in downstream analyses)
}

