% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapInput.R
\name{heatmapInput}
\alias{heatmapInput}
\title{Creates the matrix that can be easily plotted with a heatmap function available in an R package}
\usage{
heatmapInput(aggregateMatrix_withAnnotation)
}
\arguments{
\item{aggregateMatrix_withAnnotation}{the annotated clade matrix}
}
\value{
the heatmap input matrix
}
\description{
Creates the matrix that can be easily plotted with a heatmap function available in an R package
}

