% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardSM.R
\name{JaccardSM}
\alias{JaccardSM}
\title{Calculates the Jaccard similarity matrix.}
\usage{
JaccardSM(network1, network2)
}
\arguments{
\item{network1}{first network matrix input}

\item{network2}{second network matrix input}
}
\value{
the alignment/co-occurene score
}
\description{
Calculates the Jaccard similarity matrix.
}

