% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeList.R
\name{makeList}
\alias{makeList}
\title{Helper function to obtain a JSON-friendly format of a matrix formatted object}
\usage{
makeList(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A list for generating JSON format output.
}
\description{
Helper function to obtain a JSON-friendly format of a matrix formatted object
}

