% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAN.R
\name{MAN}
\alias{MAN}
\title{Creates network alignments using network constructed from subpopulations after PAC}
\usage{
MAN(sampleIDs, num_PACSupop, smallSubpopCutoff, k_clades)
}
\arguments{
\item{sampleIDs}{sampleID vector}

\item{num_PACSupop}{number of subpopulations learned in PAC step for each sample}

\item{smallSubpopCutoff}{Population size cutoff for subpopulations in clade calculation. The small subpopulations will be considered in the refinement step.}

\item{k_clades}{number of clades to output before refinement}
}
\value{
clades_network_only the clades constructed without small subpopulations (by cutoff) using mutual information network alignments
}
\description{
Creates network alignments using network constructed from subpopulations after PAC
}

