% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINetwork_matrix_topEdges.R
\name{MINetwork_matrix_topEdges}
\alias{MINetwork_matrix_topEdges}
\title{Mutual information network connection matrix generation (mrnet algorithm) using the parmigene package. Mutual information calculated with infotheo package.}
\usage{
MINetwork_matrix_topEdges(dataMatrix, threshold)
}
\arguments{
\item{dataMatrix}{data matrix}

\item{threshold}{the number of edges to draw for each subpopulation mutual information network}
}
\value{
the mutual information network connection matrix with top edges
}
\description{
Mutual information network connection matrix generation (mrnet algorithm) using the parmigene package. Mutual information calculated with infotheo package.
}

