% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINetwork_simpliefied_topEdges.R
\name{MINetwork_simplified_topEdges}
\alias{MINetwork_simplified_topEdges}
\title{Outputs the vectorized summary of a network based on the number of edges connected to a node}
\usage{
MINetwork_simplified_topEdges(dataMatrix, threshold)
}
\arguments{
\item{dataMatrix}{data matrix}

\item{threshold}{the number of edges to draw for each subpopulation mutual information network}
}
\description{
Outputs the vectorized summary of a network based on the number of edges connected to a node
}

