% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINetworkPlot_topEdges.R
\name{MINetworkPlot_topEdges}
\alias{MINetworkPlot_topEdges}
\title{Plots mutual information network (mrnet algorithm) connection using the parmigene package. Mutual information calculated with infotheo package.}
\usage{
MINetworkPlot_topEdges(dataMatrix, threshold)
}
\arguments{
\item{dataMatrix}{data matrix with first column being the sample ID}

\item{threshold}{the number of edges to draw for each subpopulation mutual information network}
}
\description{
Plots mutual information network (mrnet algorithm) connection using the parmigene package. Mutual information calculated with infotheo package.
}

