% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputNetworks_topEdges_matrix.R
\name{outputNetworks_topEdges_matrix}
\alias{outputNetworks_topEdges_matrix}
\title{Wrapper to output the mutual information networks for subpopulations with size larger than a desired threshold.}
\usage{
outputNetworks_topEdges_matrix(dataMatrix, subpopulationLabels, threshold)
}
\arguments{
\item{dataMatrix}{data matrix with first column being the sample ID}

\item{subpopulationLabels}{the subpopulation labels}

\item{threshold}{the number of edges to draw for each subpopulation mutual information network}
}
\description{
Wrapper to output the mutual information networks for subpopulations with size larger than a desired threshold.
}

