% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputRepresentativeNetworks_topEdges.R
\name{outputRepresentativeNetworks_topEdges}
\alias{outputRepresentativeNetworks_topEdges}
\title{Outputs the representative/clade networks (plots and summary vectors) for subpopulations with size larger than a desired threshold. Saves the networks and the data matrices without the smaller subpopulations.}
\usage{
outputRepresentativeNetworks_topEdges(dataMatrix, subpopulationLabels,
  threshold)
}
\arguments{
\item{dataMatrix}{data matrix with first column being the sample ID}

\item{subpopulationLabels}{the subpopulation labels}

\item{threshold}{the number of edges to draw for each subpopulation mutual information network}
}
\description{
Outputs the representative/clade networks (plots and summary vectors) for subpopulations with size larger than a desired threshold. Saves the networks and the data matrices without the smaller subpopulations.
}

