% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packedCircleInput.R
\name{packedCircleInput}
\alias{packedCircleInput}
\title{Make packed circles plot illustrating the cluster proportions across multiple samples.}
\usage{
packedCircleInput(data_agg, cluster_names, CountFilter = 1000,
  filename = "zoomablePackedCirclesInput.txt")
}
\arguments{
\item{data_agg}{Variable name of aggregated data.}

\item{cluster_names}{The annotated cluster names}

\item{CountFilter}{The filter threshold to remove clusters with event numbers below the threshold.}

\item{filename}{Set to zoomablePackedCirclesInput.txt or the name of choice for the html that draws the packed circles.}
}
\value{
Saves the D3 packed circles plot input in the working directory.
}
\description{
Make packed circles plot illustrating the cluster proportions across multiple samples.
}

