% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordWithinClusterSpread.R
\name{recordWithinClusterSpread}
\alias{recordWithinClusterSpread}
\title{Calculates the within cluster spread}
\usage{
recordWithinClusterSpread(sampleIDs, dim_subset = NULL, SubpopSizeFilter)
}
\arguments{
\item{sampleIDs}{A vector of sample names.}

\item{dim_subset}{a string vector of string names to subset the data columns for PAC; set to NULL to use all columns.}

\item{SubpopSizeFilter}{threshold to filter out very small clusters with too few points; these very small subpopulations may not be outliers and not biologically relevant.}
}
\value{
Returns the sample within cluster spread
}
\description{
Calculates the within cluster spread
}

