% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineSubpopulationLabels.R
\name{refineSubpopulationLabels}
\alias{refineSubpopulationLabels}
\title{Refines the subpopulation labels from PAC using network alignment and small subpopulation information. Outputs a new set of files containing the representative labels.}
\usage{
refineSubpopulationLabels(sampleIDs, clades_network_only, expressionGroupClamp)
}
\arguments{
\item{sampleIDs}{sampleID vector}

\item{clades_network_only}{the alignment results from MAN; used to translate the original sample-specific labels into clade labels}

\item{expressionGroupClamp}{clamps the subpopulations into desired number of expression groups for assigning small subpopulations into larger groups or their own groups.}
}
\description{
Refines the subpopulation labels from PAC using network alignment and small subpopulation information. Outputs a new set of files containing the representative labels.
}

