% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamePrunedSubpopulations.R
\name{renamePrunedSubpopulations}
\alias{renamePrunedSubpopulations}
\title{Prune away specified subpopulations in clades that are far away.}
\usage{
renamePrunedSubpopulations(pacman_results, subpopulationsToPrune)
}
\arguments{
\item{pacman_results}{PAC-MAN analysis result matrix that contains network annotation, clade IDs and mean (centroid) clade expression levels.}

\item{subpopulationsToPrune}{A vector of clade IDs; these clades will be pruned.}
}
\value{
Returns PAC-MAN analysis result matrix with pruned clades. The pruning process creates new clades to replace the original clade ID of the specified subpopulations.
}
\description{
Prune away specified subpopulations in clades that are far away.
}

