% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePass.R
\name{samplePass}
\alias{samplePass}
\title{A wrapper to run PAC and output subpopulation mutual information networks.}
\usage{
samplePass(sampleIDs, hyperrectangles, num_PACSupop, max.iter, num_networkEdge)
}
\arguments{
\item{sampleIDs}{sampleID vector}

\item{hyperrectangles}{number of hyperrectangles to learn for each sample}

\item{num_PACSupop}{number of subpopulations to output for each sample using PAC}

\item{max.iter}{postprocessing kmeans iterations}

\item{num_networkEdge}{a threshold on the number of edges to output for each subpopulation mutual information network}
}
\description{
A wrapper to run PAC and output subpopulation mutual information networks.
}

