% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalLevelHeatmap.R
\name{signalLevelHeatmap}
\alias{signalLevelHeatmap}
\title{Make heatmap illustrating the signel levels for clusters. For multiple samples, signal levels differ for the same cluster in different samples. Signal levels can be plotted easily using built-in R functions such as mean, median, and max.}
\usage{
signalLevelHeatmap(data_agg, signal_level = max)
}
\arguments{
\item{data_agg}{Variable name of aggregated data.}

\item{signal_level}{Function used to find signal levels for visualization.}
}
\value{
Saves the cluster signal level heatmap plot in the working directory.
}
\description{
Make heatmap illustrating the signel levels for clusters. For multiple samples, signal levels differ for the same cluster in different samples. Signal levels can be plotted easily using built-in R functions such as mean, median, and max.
}

