% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregates results from the clustering and merging step.}
\usage{
aggregateData(dataInput, labelsInput)
}
\arguments{
\item{dataInput}{Data matrix, with first column being SampleID.}

\item{labelsInput}{cluster labels from PAC.}
}
\value{
The aggregated data of \code{dataInput}, with average signal levels for all clusters and sample combinations.
}
\description{
Aggregates results from the clustering and merging step.
}

